requirements.txt
diffusers==0.29.2
imageio==2.21.1
jaxtyping==0.2.36
kaolin==0.17.0
lib==4.0.0
matplotlib==3.9.4
moviepy==1.0.3
numpy==1.23.1
nvdiffrast @ git+https://github.com/NVlabs/nvdiffrast.git
omegaconf==2.3.0
open3d==0.18.0
opencv_contrib_python==4.10.0.84
opencv_python==4.6.0.66
opencv_python_headless==4.10.0.84
packaging==24.2
Pillow==9.5.0
Pillow==11.0.0
pymeshlab==2023.12.post2
pytorch3d==0.7.4
rembg==0+untagged.232.g5f4386b
scipy==1.14.1
termcolor==2.5.0
torch==2.0.1+cu118
torch_scatter==2.1.2
torchvision==0.15.2+cu118
tqdm==4.67.0
tqdm==4.64.1
transformers==4.30.2
trimesh==4.4.1
typeguard==4.4.1
typeguard==4.3.0
yacs==0.1.8
